requirements.txt
pytest==7.3.0
selene==2.0.0rc3post2
allure-pytest==2.13.2
selenium==4.11.2
webdriver-manager==3.8.6
python-dotenv
